# specs/001-mediawatch-saas-platform/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: MediaWatch CI API
  description: |
    API REST pour la plateforme MediaWatch CI - Veille médias automatisée pour la Côte d'Ivoire.
    
    Cette API permet aux clients de gérer leurs mots-clés, consulter leurs mentions,
    configurer des alertes, et accéder aux analytics de veille médias.
  version: 1.0.0
  contact:
    name: MediaWatch CI Support
    email: support@mediawatch.ci
servers:
  - url: https://api.mediawatch.ci/v1
    description: Production
  - url: https://staging-api.mediawatch.ci/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Développement local
...
security:
  - BearerAuth: []
paths:
...
  /keywords:
    get:
      tags: [keywords]
      summary: Lister les mots-clés
      description: Récupère tous les mots-clés de l'organisation de l'utilisateur
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filtrer par statut activé/désactivé
      responses:
        '200':
          description: Liste des mots-clés
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Keyword'
                  total:
                    type: integer
                  limit_reached:
                    type: boolean
                    description: True si limite de plan atteinte

    post:
      tags: [keywords]
      summary: Créer un mot-clé
      description: Ajoute un nouveau mot-clé à surveiller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: Orange CI
                category:
                  type: string
                  enum: [brand, product, person, competitor, custom]
                  example: brand
                alert_enabled:
                  type: boolean
                  default: true
                alert_threshold:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                  default: 0.3
                  description: Seuil de sentiment négatif pour alertes (0-1)
      responses:
        '201':
          description: Mot-clé créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Mot-clé déjà existant ou limite atteinte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /mentions:
    get:
      tags: [mentions]
      summary: Lister les mentions
      description: Récupère les mentions avec filtres et pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            enum: [25, 50, 100]
            default: 25
        - name: keyword_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrer par mot-clé
        - name: source_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrer par source
        - name: sentiment
          in: query
          schema:
            type: string
            enum: [negative, neutral, positive]
          description: Filtrer par sentiment
        - name: theme
          in: query
          schema:
            type: string
            enum: [politics, economy, sport, society]
          description: Filtrer par thème
        - name: date_from
          in: query
          schema:
            type: string
            format: date
          description: Date début (YYYY-MM-DD)
        - name: date_to
          in: query
          schema:
            type: string
            format: date
          description: Date fin (YYYY-MM-DD)
        - name: sort
          in: query
          schema:
            type: string
            enum: [date_desc, date_asc, sentiment_desc, visibility_desc]
            default: date_desc
      responses:
        '200':
          description: Liste des mentions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mention'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /admin/sources:
    get:
      tags: [admin]
      summary: Lister toutes les sources (Admin)
      description: Récupère toutes les sources avec statut de scraping
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ok, failing, disabled]
      responses:
        '200':
          description: Liste des sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceAdmin'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/sources/{source_id}/retry:
    post:
      tags: [admin]
      summary: Réessayer scraping source (Admin)
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Scraping relancé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /admin/organizations/{org_id}/suspend:
    post:
      tags: [admin]
      summary: Suspendre organisation (Admin)
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Organisation suspendue
  /admin/organizations/{org_id}/reactivate:
    post:
      tags: [admin]
      summary: Réactiver organisation (Admin)
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organisation réactivée

# ==================== COMPONENTS ====================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Keyword:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        category:
          type: string
        enabled:
          type: boolean
        alert_enabled:
          type: boolean
        alert_threshold:
          type: number
          format: float
        total_mentions_count:
          type: integer
        last_mention_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    Mention:
      type: object
      properties:
        id:
          type: string
          format: uuid
        keyword:
          type: object
          properties:
            id:
              type: string
              format: uuid
            text:
              type: string
        article:
          type: object
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            url:
              type: string
              format: uri
            published_at:
              type: string
              format: date-time
        source:
          $ref: '#/components/schemas/Source'
        matched_text:
          type: string
        match_context:
          type: string
          description: Extrait avec contexte autour du match
        sentiment_score:
          type: number
          format: float
          minimum: -1
          maximum: 1
        sentiment_label:
          type: string
          enum: [negative, neutral, positive]
        visibility_score:
          type: number
          format: float
        theme:
          type: string
        detected_at:
          type: string
          format: date-time
...
    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [press, whatsapp, rss, api]
        prestige_score:
          type: number
          format: float
    SourceAdmin:
      allOf:
        - $ref: '#/components/schemas/Source'
        - type: object
          properties:
            scraping_enabled:
              type: boolean
            last_scrape_at:
              type: string
              format: date-time
            last_success_at:
              type: string
              format: date-time
            consecutive_failures:
              type: integer
            last_error_message:
              type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Accès refusé
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...